+++ packages/chainkey_ios/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+* Map platform exceptions to typed `ChainkeyException` instances.
+* Integrate updated `chainkey_platform_interface` dependency.
+
 ## 0.1.0
 
 * Initial release of chainkey federated plugin package.